replace pre-refactor scripts
